+++ R/io.R
@@ -74,7 +74,7 @@ read_utf8_bare <- function(con, warn = TRUE) {
         "The file ", con, " is not encoded in UTF-8. ",
         "These lines contain invalid UTF-8 characters: "
       ),
-      paste(c(head(i), if (n > 6) "..."), collapse = ", ")
+      paste(c(utils::head(i), if (n > 6) "..."), collapse = ", ")
     )
   }
   x